+++ splash.c
@@ -9,31 +9,31 @@
 CP_Image logo;
 BYTE opacity = 0;
 
-float elapsedTime = 0;
-float elapsedShow = 0;
+double elapsedTime = 0;
+double elapsedShow = 0;
 
 enum { FADE_IN, SHOW_LOGO, FADE_OUT };
 BYTE splashState = FADE_IN;
 
 
 void fadeIn() {
-	float dt = CP_System_GetDt();
+	double dt = CP_System_GetDt();
 	elapsedTime += dt;
-	float opacityPercentage = elapsedTime / FADE_DURATION;
+	double opacityPercentage = elapsedTime / FADE_DURATION;
 	opacityPercentage = min(1, opacityPercentage);
 	opacity = (BYTE)(255 * opacityPercentage);
 }
 
 void fadeOut() {
-	float dt = CP_System_GetDt();
+	double dt = CP_System_GetDt();
 	elapsedTime -= dt;
-	float opacityPercentage = elapsedTime / FADE_DURATION;
+	double opacityPercentage = elapsedTime / FADE_DURATION;
 	opacityPercentage = max(0, opacityPercentage);
 	opacity = (BYTE)(255 * opacityPercentage);
 }
 
 void splashScreenFadeInOut() {
-	float dt = CP_System_GetDt();
+	double dt = CP_System_GetDt();
 	
 	if (opacity == 0 && splashState == FADE_OUT) {
 			CP_Engine_SetNextGameState(menuInit, menuUpdate, menuExit);
@@ -60,7 +60,7 @@ void splashScreenFadeInOut() {
 			break;
 	}
 
-	CP_Image_Draw(logo, WINDOW_SIZE.width / 2, WINDOW_SIZE.height / 2, (float)CP_Image_GetWidth(logo), (float)CP_Image_GetHeight(logo), opacity);
+	CP_Image_Draw(logo, (float)(WINDOW_SIZE.width / 2), (float)(WINDOW_SIZE.height / 2), (float)CP_Image_GetWidth(logo), (float)CP_Image_GetHeight(logo), opacity);
 }
 
 void splash_screen_init(void) {
